move collection stuff to its own file
