+++ fast64_internal/oot/cutscene/properties.py
@@ -2,7 +2,8 @@ from bpy.types import PropertyGroup, Object, UILayout, Scene, Context
 from bpy.props import StringProperty, EnumProperty, IntProperty, BoolProperty, CollectionProperty, PointerProperty
 from bpy.utils import register_class, unregister_class
 from ...utility import PluginError, prop_split
-from ..oot_utility import OOTCollectionAdd, drawCollectionOps, getEnumName
+from ..collection_utility import OOTCollectionAdd, drawCollectionOps
+from ..oot_utility import getEnumName
 from ..oot_constants import ootData
 from ..oot_upgrade import upgradeCutsceneSubProps, upgradeCSListProps, upgradeCutsceneProperty
 from .operators import OOTCSTextAdd, OOT_SearchCSDestinationEnumOperator, OOTCSListAdd, OOT_SearchCSSeqOperator
