+++ fast64_internal/oot/oot_utility.py
@@ -6,8 +6,6 @@ import re
 from ast import parse, Expression, Num, UnaryOp, USub, Invert, BinOp
 from mathutils import Vector
 from bpy.types import Object
-from bpy.utils import register_class, unregister_class
-from bpy.types import Object
 from typing import Callable, Optional, TYPE_CHECKING, List
 from .oot_constants import ootSceneIDToName
 from dataclasses import dataclass
@@ -21,7 +19,6 @@ from ..utility import (
     setOrigin,
     applyRotation,
     cleanupDuplicatedObjects,
-    ootGetSceneOrRoomHeader,
     hexOrDecInt,
     binOps,
 )
@@ -645,160 +642,6 @@ def getCutsceneName(obj):
     return name
 
 
-def getCollectionFromIndex(obj, prop, subIndex, isRoom):
-    header = ootGetSceneOrRoomHeader(obj, subIndex, isRoom)
-    return getattr(header, prop)
-
-
-# Operators cannot store mutable references (?), so to reuse PropertyCollection modification code we do this.
-# Save a string identifier in the operator, then choose the member variable based on that.
-# subIndex is for a collection within a collection element
-def getCollection(objName, collectionType, subIndex):
-    obj = bpy.data.objects[objName]
-    if collectionType == "Actor":
-        collection = obj.ootActorProperty.headerSettings.cutsceneHeaders
-    elif collectionType == "Transition Actor":
-        collection = obj.ootTransitionActorProperty.actor.headerSettings.cutsceneHeaders
-    elif collectionType == "Entrance":
-        collection = obj.ootEntranceProperty.actor.headerSettings.cutsceneHeaders
-    elif collectionType == "Room":
-        collection = obj.ootAlternateRoomHeaders.cutsceneHeaders
-    elif collectionType == "Scene":
-        collection = obj.ootAlternateSceneHeaders.cutsceneHeaders
-    elif collectionType == "Light":
-        collection = getCollectionFromIndex(obj, "lightList", subIndex, False)
-    elif collectionType == "Exit":
-        collection = getCollectionFromIndex(obj, "exitList", subIndex, False)
-    elif collectionType == "Object":
-        collection = getCollectionFromIndex(obj, "objectList", subIndex, True)
-    elif collectionType == "Curve":
-        collection = obj.ootSplineProperty.headerSettings.cutsceneHeaders
-    elif collectionType.startswith("CSHdr."):
-        # CSHdr.HeaderNumber[.ListType]
-        # Specifying ListType means uses subIndex
-        toks = collectionType.split(".")
-        assert len(toks) in [2, 3]
-        hdrnum = int(toks[1])
-        collection = getCollectionFromIndex(obj, "csLists", hdrnum, False)
-        if len(toks) == 3:
-            collection = getattr(collection[subIndex], toks[2])
-    elif collectionType.startswith("Cutscene."):
-        # Cutscene.ListType
-        toks = collectionType.split(".")
-        assert len(toks) == 2
-        collection = obj.ootCutsceneProperty.csLists
-        collection = getattr(collection[subIndex], toks[1])
-    elif collectionType == "Cutscene":
-        collection = obj.ootCutsceneProperty.csLists
-    elif collectionType == "extraCutscenes":
-        collection = obj.ootSceneHeader.extraCutscenes
-    elif collectionType == "BgImage":
-        collection = obj.ootRoomHeader.bgImageList
-    else:
-        raise PluginError("Invalid collection type: " + collectionType)
-
-    return collection
-
-
-def drawAddButton(layout, index, collectionType, subIndex, objName):
-    if subIndex is None:
-        subIndex = 0
-    addOp = layout.operator(OOTCollectionAdd.bl_idname)
-    addOp.option = index
-    addOp.collectionType = collectionType
-    addOp.subIndex = subIndex
-    addOp.objName = objName
-
-
-def drawCollectionOps(layout, index, collectionType, subIndex, objName, allowAdd=True, compact=False):
-    if subIndex is None:
-        subIndex = 0
-
-    if not compact:
-        buttons = layout.row(align=True)
-    else:
-        buttons = layout
-
-    if allowAdd:
-        addOp = buttons.operator(OOTCollectionAdd.bl_idname, text="Add" if not compact else "", icon="ADD")
-        addOp.option = index + 1
-        addOp.collectionType = collectionType
-        addOp.subIndex = subIndex
-        addOp.objName = objName
-
-    removeOp = buttons.operator(OOTCollectionRemove.bl_idname, text="Delete" if not compact else "", icon="REMOVE")
-    removeOp.option = index
-    removeOp.collectionType = collectionType
-    removeOp.subIndex = subIndex
-    removeOp.objName = objName
-
-    moveUp = buttons.operator(OOTCollectionMove.bl_idname, text="Up" if not compact else "", icon="TRIA_UP")
-    moveUp.option = index
-    moveUp.offset = -1
-    moveUp.collectionType = collectionType
-    moveUp.subIndex = subIndex
-    moveUp.objName = objName
-
-    moveDown = buttons.operator(OOTCollectionMove.bl_idname, text="Down" if not compact else "", icon="TRIA_DOWN")
-    moveDown.option = index
-    moveDown.offset = 1
-    moveDown.collectionType = collectionType
-    moveDown.subIndex = subIndex
-    moveDown.objName = objName
-
-
-class OOTCollectionAdd(bpy.types.Operator):
-    bl_idname = "object.oot_collection_add"
-    bl_label = "Add Item"
-    bl_options = {"REGISTER", "UNDO"}
-
-    option: bpy.props.IntProperty()
-    collectionType: bpy.props.StringProperty(default="Actor")
-    subIndex: bpy.props.IntProperty(default=0)
-    objName: bpy.props.StringProperty()
-
-    def execute(self, context):
-        collection = getCollection(self.objName, self.collectionType, self.subIndex)
-
-        collection.add()
-        collection.move(len(collection) - 1, self.option)
-        return {"FINISHED"}
-
-
-class OOTCollectionRemove(bpy.types.Operator):
-    bl_idname = "object.oot_collection_remove"
-    bl_label = "Remove Item"
-    bl_options = {"REGISTER", "UNDO"}
-
-    option: bpy.props.IntProperty()
-    collectionType: bpy.props.StringProperty(default="Actor")
-    subIndex: bpy.props.IntProperty(default=0)
-    objName: bpy.props.StringProperty()
-
-    def execute(self, context):
-        collection = getCollection(self.objName, self.collectionType, self.subIndex)
-        collection.remove(self.option)
-        return {"FINISHED"}
-
-
-class OOTCollectionMove(bpy.types.Operator):
-    bl_idname = "object.oot_collection_move"
-    bl_label = "Move Item"
-    bl_options = {"REGISTER", "UNDO"}
-
-    option: bpy.props.IntProperty()
-    offset: bpy.props.IntProperty()
-    subIndex: bpy.props.IntProperty(default=0)
-    objName: bpy.props.StringProperty()
-
-    collectionType: bpy.props.StringProperty(default="Actor")
-
-    def execute(self, context):
-        collection = getCollection(self.objName, self.collectionType, self.subIndex)
-        collection.move(self.option, self.option + self.offset)
-        return {"FINISHED"}
-
-
 def getHeaderSettings(actorObj: bpy.types.Object):
     itemType = actorObj.ootEmptyType
     if actorObj.type == "EMPTY":
@@ -816,23 +659,6 @@ def getHeaderSettings(actorObj: bpy.types.Object):
         headerSettings = None
 
     return headerSettings
-
-
-oot_utility_classes = (
-    OOTCollectionAdd,
-    OOTCollectionRemove,
-    OOTCollectionMove,
-)
-
-
-def oot_utility_register():
-    for cls in oot_utility_classes:
-        register_class(cls)
-
-
-def oot_utility_unregister():
-    for cls in reversed(oot_utility_classes):
-        unregister_class(cls)
 
 
 def getActiveHeaderIndex() -> int:
