+++ fast64_internal/oot/room/properties.py
@@ -2,7 +2,8 @@ import bpy
 from bpy.types import PropertyGroup, UILayout, Image, Object
 from bpy.utils import register_class, unregister_class
 from ...utility import prop_split
-from ..oot_utility import drawCollectionOps, onMenuTabChange, onHeaderMenuTabChange, drawEnumWithCustom, drawAddButton
+from ..collection_utility import drawCollectionOps, drawAddButton
+from ..oot_utility import onMenuTabChange, onHeaderMenuTabChange, drawEnumWithCustom
 from ..oot_upgrade import upgradeRoomHeaders
 from .operators import OOT_SearchObjectEnumOperator
